output/dnssim: refactor - split udp dnssim into separate file
